Create cronjob.yml
What This YAML File Does

Creates a Kubernetes CronJob:
The CronJob is named myjob and is set to run every 3 minutes based on the specified cron expression.

Job Execution:
Each time the CronJob runs, a job will be created based on the jobTemplate. The job runs a container with the image luksa/batch-job.

Scheduled Execution:
The CronJob is scheduled to execute based on the cron expression "*/3 * * * *", meaning it will run every 3 minutes.

Pod Template:
The pod created for each job will have a label app: myapp for easier identification.
The pod will have a restartPolicy of OnFailure, so it will only restart if the container fails.

Container Definition:
Each job pod will run a container (mycont) using the luksa/batch-job image, which likely contains the logic for the batch job.